# Conflicts:
#	libraries/remote_config/src/main/java/com/tokopedia/remoteconfig/RemoteConfigKey.java
#	sellerapp/src/main/java/com/tokopedia/sellerapp/SellerRouterApplication.java

+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/domain/model/response/atcexternal/AddToCartExternalDataModel.kt
@@ -7,7 +7,7 @@ data class AddToCartExternalModel(
 )
 
 data class AddToCartExternalDataModel(
-        var productId: Int = 0,
+        var productId: Long = 0,
         var productName: String = "",
         var quantity: Int = 0,
         var price: Int = 0,
